move search method from main wrapper component to app component
