+++ src/components/App.js
@@ -4,6 +4,7 @@ import Warnings from './Warnings';
 import Top from './Top/Top';
 import Footer from './Footer/Footer';
 import {getAddressBalance} from '../lib/dAppService';
+import {entries} from '../lib/ensService';
 import 'typeface-roboto';
 import './App.css';
 
@@ -16,14 +17,65 @@ class App extends Component {
       source: 'keystore',
       page: 'main',
       keystore: '',
-      privateKey: ''
+      privateKey: '',
+      
+      // for handling search
+      searchValue: '',
+      searchResult: null,
+      searchFetching: false,
+      warningMessage: '',
+      warningOpen: false,
     };
+
     this.setAddress = this.setAddress.bind(this);
     this.setSource = this.setSource.bind(this);
     this.switchPage = this.switchPage.bind(this);
     this.setKeystore = this.setKeystore.bind(this);
     this.setPrivateKey = this.setPrivateKey.bind(this);
+    
+    // for handling search
+    this.handleSearchChange = this.handleSearchChange.bind(this);
+    this.handleSearchKeyPress = this.handleSearchKeyPress.bind(this);
+    this.handleSearchClick = this.handleSearchClick.bind(this);
+    this.handleMessageOpen = this.handleMessageOpen.bind(this);
+    this.handleMessageClose = this.handleMessageClose.bind(this);
+    this.backToSearch = this.backToSearch.bind(this);
   }
+
+  handleSearchChange(e) {
+    this.setState({searchValue: e.target.value});
+  }
+
+  handleSearchKeyPress(e) {
+    if (e.key === 'Enter') {
+      this.handleSearchClick(e);
+    }
+  }
+
+  handleSearchClick(e) {
+    e.preventDefault();
+    if ((this.state.searchValue).length < 7) {
+      this.handleMessageOpen('ENS .eth should greater than 7 words');
+    } else {
+      if (this.state.searchValue) {
+        this.setState({fetching: true}); //TODO: not work
+        const searchResult = entries(this.state.searchValue);
+        searchResult.searchName = this.state.searchValue;
+        this.setState({searchResult, searchFetching: false});
+      }
+    }
+  }
+
+  backToSearch() {
+    this.switchPage('main');
+    this.setState({
+      searchValue: null,
+      searchResult: null
+    });
+  }
+
+  handleMessageOpen = msg => this.setState({warningOpen: true, warningMessage: msg});
+  handleMessageClose = () => this.setState({warningOpen: false});
 
   setSource(type) {
     this.setState({source: type});
@@ -52,6 +104,12 @@ class App extends Component {
           className="App-content"
           {...this.state}
           switchPage={this.switchPage}
+          handleSearchChange={this.handleSearchChange}
+          handleSearchClick={this.handleSearchClick}
+          handleSearchKeyPress={this.handleSearchKeyPress}
+          backToSearch={this.backToSearch}
+          handleMessageOpen={this.handleMessageOpen}
+          handleMessageClose={this.handleMessageClose}
         />
       : <Warnings/>;
     
@@ -63,7 +121,7 @@ class App extends Component {
           setSource={this.setSource}
           setKeystore={this.setKeystore}
           setPrivateKey={this.setPrivateKey}
-          switchPage={this.switchPage}
+          backToSearch={this.backToSearch}
         /> 
         {mainWrapperOrWarning}
         <Footer/>
